Fix edge case in exception handler causing recursion
During Plone's porting efforts to Python 3.6 a bug was found
that caused Python to dump core with a stack overflow. This
is caused in the field handling, where a template calls a
macro which then inserts the rendered body of another pagetemplate.

The inner page template is rendered and calls the exception handler
correctly, which causes the exception's `__str__` method to be
replaced with an instance of `ExceptionFormatter`. However, as
this is within a macro the `visit_Macro` function causes that
exception to be added to `rcontext['__error__']``, meaning that
when the outer template attempts to format its exception the
handling of the `__errors__` list causes recursion into the
`ExceptionFormatter` until the maximum recursion depth is reached
at which point the error handling is aborted as unformattable
by the traceback module in the standard library.

If the exception is sufficiently large it smashes the stack before
this can happen, causing a core dump. If it isn't sufficiently
large a confusingly repetitious error is shown.