Changes to string interpolation dollar escape coding
Fix double dollar '$$' escaping such that a double dollar is always
resolved, either as an interpolation expression, or as an escape
where it is substituted by a single dollar symbol. This is now
consistent with Zope's handling of this character.

Backslash-escaping of dollar-based string interpolation is no longer
supported. The documentation has been updated to reflect this
change.