Avoid various traceback reference cycles
In Python 3, exceptions have a ``__traceback__`` attribute containing
their associated traceback.  Storing an exception in a local variable of
a frame present in that traceback thus creates a reference cycle.  Avoid
this either by explicitly deleting the exception value when we're
finished with it, or by using the simpler ``except ... as`` construction
(since Python 3 automatically clears the exception at the end of the
``except`` clause in that case).